Add files via upload
this is my open ended lab task to run all the models which is studied so further and see which model is better. After performing all the functions it is said that KNN is better in predicting price rather then linear regression. The  accuracy is slightly different of both models. As it is our first open ended so I work for a long time on this lab as it is my first implementation.Thankyou :)